==> 02_remove_overlap_wyscout_statsbomb.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from mplsoccer.pitch import Pitch

Data Folders

WYSCOUT = os.path.join('data', 'wyscout')
STATSBOMB = os.path.join('data', 'statsbomb')

Load match data

df_wyscout_match = pd.read_parquet(os.path.join(WYSCOUT, 'match.parquet'))
df_statsbomb_match = pd.read_parquet(os.path.join(STATSBOMB, 'match.parquet'))

Overlaps

match_sb = df_statsbomb_match[['match_id', 'home_team_name', 'away_team_name', 'kick_off', 'competition_name']]
match_wy = df_wyscout_match[['match_id', 'home_team_name', 'away_team_name', 'kick_off' ,'competition_name']]
overlap = match_wy.merge(match_sb, on=['home_team_name', 'away_team_name', 'kick_off', 'competition_name'],
                         how='inner', suffixes=['_wyscout', '_statsbomb'])

There are 100 overlapping games. 64 world cup games and 36 La Liga games from the 2017/18 season. Barcelona actually played 38 La Liga games in 2017/18, but Messi did not play 2 of those games.

overlap.competition_name.value_counts()

Remove overlapping games from Wyscout data

df_wyscout_event = pd.read_parquet(os.path.join(WYSCOUT, 'event.parquet'))

mask = df_wyscout_event.match_id.isin(overlap.match_id_wyscout)
df_wyscout_event_overlap = df_wyscout_event[mask].copy()
df_wyscout_event = df_wyscout_event[~mask].copy()
print('Number of events removed:', mask.sum())

Save events

df_wyscout_event_overlap.to_parquet(os.path.join(WYSCOUT, 'event_overlap.parquet'))
df_wyscout_event.to_parquet(os.path.join(WYSCOUT, 'event.parquet'))
